PP-13337 Patch our TBB fork debug build on Windows
This is intended to be a temporary workaround and
should be properly fixed by
https://formlabs.atlassian.net/browse/PP-13395 as
soon as possible.

Note that https://github.com/Formlabs/tbb is a
submodule in https://github.com/Formlabs/euc which
is a submodule in
https://github.com/Formlabs/PreForm/tree/master/vendor.
All 3 repos must be updated to have this patch.

Again, this is intended to be a very nasty and very
temporary hack to enable debug builds on Windows but
needs to be fixed by
https://formlabs.atlassian.net/browse/PP-13395.